requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-Login==0.6.3
python-socketio==5.10.0
python-engineio==4.8.0
eventlet==0.33.3
bcrypt==4.1.1
websocket-client==1.7.0
